prompt_eng/agents/vector_store.py
from typing import List
from document import Document
import numpy as np
import faiss

class VectorStore:
    def __init__(self):
        self.index = None
        self.documents = []
        
    def add_documents(self, documents: List[Document]):
        # Implement FAISS index creation/updating
        pass
        
    def search(self, query: str, top_k: int = 3):
        # Implement similarity search
        return self.documents[:top_k]  # Temporary placeholder 