prompt_eng/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
asyncio>=3.4.3
typing>=3.7.4.3
requests>=2.31.0
streamlit>=1.32.0
faiss-cpu>=1.7.4  # For vector store
numpy>=1.21.0
aiohttp>=3.9.0  # For async HTTP requests
pydantic>=2.6.0  # For data validation
jinja2>=3.1.0  # For code template rendering
python-dateutil>=2.8.2  # For date handling
pytz>=2024.1  # For timezone support